Add search functionality and sort question bases on helpfulness
