+++ client/src/components/QnA/QuestionsNAnswersContainer.jsx
@@ -1,6 +1,5 @@
 import React from 'react';
 import SearchQuestions from './SearchQuestions.jsx';
-import QuestionsList from './QuestionsList.jsx';
 import './QnA.css';
 
 class QuestionsNAnswersContainer extends React.Component {
@@ -11,8 +10,7 @@ class QuestionsNAnswersContainer extends React.Component {
   render() {
     return (
       <div className="questions-answers-container">
-        <SearchQuestions />
-        <QuestionsList formatBody={this.props.formatBody}/>
+        <SearchQuestions formatBody={this.props.formatBody} />
       </div>
     );
   }
